now the cost is returned as double. (fix infinite loop bug when evaluating same cost solutions)

+++ src/it/polimi/modaclouds/space4cloud/optimization/solution/impl/Solution.java
@@ -433,8 +433,8 @@ public class Solution implements Cloneable, Serializable {
 	 * 
 	 * @return the cost
 	 */
-	public float getCost() {
-		return (float)cost;
+	public double getCost() {
+		return cost;
 	}
 
 	/**
